#6 Design : 타이머 디자인 수정
